backend/requirements.txt
# Frameworks
fastapi[all]

# Seguridad y autenticación
passlib
bcrypt
pyjwt
python-jose

# Utilidades
python-dotenv
email-validator
requests
werkzeug
pytz

# Base de datos
sqlalchemy
mysql-connector-python
aiomysql
pymongo

# Caché
fastapi-cache2
redis[asyncio]

# Para Excel
pandas==2.1.4
openpyxl==3.1.2

# Para PDF
fpdf2==2.7.9

# Validación y tipos
pydantic
starlette

